Move EjpLink to a ui atom as the same styling is to be used on redirect page
