+++ app/components/ui/atoms/FooterText.md
@@ -1,0 +1,7 @@
+Wrapper for styling footer text such as ejp link and T&C text.
+
+```js
+;<FooterText>
+  Please check our <a href="#">terms and conditions</a>
+</FooterText>
+```
